+++ include/fastscapelib/bedrock_channel.hpp
@@ -11,7 +11,7 @@
 #include "xtensor/xtensor.hpp"
 #include "xtensor/xmanipulation.hpp"
 
-#include "fastscapelib/utils.hpp"
+#include "fastscapelib/utils/utils.hpp"
 
 
 namespace fastscapelib
