+++ include/fastscapelib/utils/iterators.hpp
@@ -1,5 +1,5 @@
-#ifndef FASTSCAPELIB_ITERATORS_H
-#define FASTSCAPELIB_ITERATORS_H
+#ifndef FASTSCAPELIB_UTILS_ITERATORS_H
+#define FASTSCAPELIB_UTILS_ITERATORS_H
 
 #include "xtl/xiterator_base.hpp"
