+++ test/test_flow_routing.cpp
@@ -1,4 +1,4 @@
-#include "fastscapelib/utils.hpp"
+#include "fastscapelib/utils/utils.hpp"
 #include "fastscapelib/grid/raster_grid.hpp"
 #include "fastscapelib/flow_routing.hpp"
